Add more icon sizes
The 48x48 shows up in the extensions page, and the 16x16 can show up as the
favicon of any tabs the extension opens and the toolbar icon for the extension.
The 32x32 is apparently used in Windows some of the time, for something...

Resizing them ourselves yields _much_ crisper results than allowing Webkit to do
it for us.