Fall back to a different attribute for blackouts
Apparently, some people don't have the `reg-dispatch-modal` property on the edit
link. In that case, we fall back to an attribute @Shakathesaint found called
`a-modal`.

If neither of those exist, we no longer explode and instead just ignore blackout
parsing entirely. It's not important enough to explode over, after all.