Use a loading spinner when more items remain
We can detect when the user hasn't scrolled all the way through the
items list by the absence of the "end of list" element.

Until that element shows up, we now display a loading spinner to
indicate that more content is coming and the total isn't complete yet.

Ideally we'd include some loading text on hover, but the spinner is good
enough for now.